pyproject.toml
[project]
name = "knowledge-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "openai>=2.30.0",
    "pdf2image>=1.17.0",
    "pillow>=12.2.0",
    "pydantic-settings>=2.13.1",
    "pyobjc-framework-vision>=12.1",
    "rich>=14.3.3",
    "typer>=0.24.1",
]
